@@ -1,6 +1,6 @@
 {
   "name": "sweetalert2",
-  "version": "11.22.0",
+  "version": "11.22.2",
   "repository": "sweetalert2/sweetalert2",
   "homepage": "https://sweetalert2.github.io/",
   "description": "A beautiful, responsive, customizable and accessible (WAI-ARIA) replacement for JavaScript's popup boxes, supported fork of sweetalert",
@@ -17,8 +17,8 @@
     "@sweetalert2/stylelint-config": "^3.0.0",
     "cypress": "^14.0.0",
     "eslint": "^9.0.0",
-    "eslint-plugin-cypress": "^4.0.0",
-    "eslint-plugin-jsdoc": "^50.0.0",
+    "eslint-plugin-cypress": "^5.0.0",
+    "eslint-plugin-jsdoc": "^51.0.0",
     "eslint-plugin-no-unsanitized": "^4.0.1",
     "jquery": "^3.6.1",
     "playwright-webkit": "^1.37.1",
@@ -31,7 +31,7 @@
     "zx": "^8.0.0"
   },
   "devDependenciesComments": {
-    "typescript": "Do not upgrade TS, tests will verify that types work with the very first version of TS >= 4.7.4"
+    "typescript": "Do not upgrade TS, tests will verify that types work with the very first version of TS >= 4.8.4"
   },
   "files": [
     "dist",
