@@ -36,10 +36,10 @@ module.exports = {
 
   updateComment: async (req, res) => {
     try {
-      const { commentId } = req.params;
+      const { id } = req.params;
       const { content } = req.body;
       const comment = await commentService.updateComment(
-        commentId, 
+        id, 
         req.user.id, 
         req.user.role,
         content
@@ -62,8 +62,8 @@ module.exports = {
 
   deleteComment: async (req, res) => {
     try {
-      const { commentId } = req.params;
-      await commentService.deleteComment(commentId, req.user.id, req.user.role);
+      const { id } = req.params;
+      await commentService.deleteComment(id, req.user.id, req.user.role);
       return responseUtils.success(res, { message: 'Comment deleted successfully' });
     } catch (error) {
       console.error('Delete comment error:', error);
@@ -93,6 +93,29 @@ module.exports = {
       return responseUtils.success(res, result);
     } catch (error) {
       console.error('Get all comments error:', error);
+      return responseUtils.serverError(res, error.message);
+    }
+  },
+
+  /**
+   * Bật/tắt bình luận cho bài post
+   * PATCH /posts/:postId/comments/enabled
+   * Body: { enabled: true/false }
+   */
+  setCommentsEnabled: async (req, res) => {
+    try {
+      const { postId } = req.params;
+      const { enabled } = req.body;
+      const post = await commentService.setCommentsEnabled(postId, enabled);
+      return responseUtils.success(res, {
+        message: `Comments have been ${enabled ? 'enabled' : 'disabled'} for this post`,
+        post
+      });
+    } catch (error) {
+      console.error('Set comments enabled error:', error);
+      if (error.message === 'Post not found') {
+        return responseUtils.notFound(res);
+      }
       return responseUtils.serverError(res, error.message);
     }
   }
