@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { ActivatedRoute } from '@angular/router';
+import { ActivatedRoute, Router } from '@angular/router';
 import { BlogPostService } from '../../core/services/blog-post.service';
 import { Post } from '../../shared/model/post.model';
 import { CommonModule } from '@angular/common';
@@ -17,6 +17,10 @@ import { throwError, of } from 'rxjs';
 import { User } from '../../shared/model/user.model';
 import { CloudinaryService } from '../../core/services/cloudinary.service';
 import { Md5 } from 'ts-md5';
+import { DomSanitizer, SafeHtml } from '@angular/platform-browser';
+import DOMPurify from 'dompurify';
+import Swal from 'sweetalert2';
+import { LanguagesService } from '../../core/services/languages.service';
 
 interface ApiResponse<T> {
   success: boolean;
@@ -42,6 +46,7 @@ interface ApiResponse<T> {
 })
 export class PostDetailComponent implements OnInit {
   post: Post | null = null;
+  originalPost: Post | null = null; // Lưu post gốc
   comments: Comment[] = [];
   loading = true;
   error: string | null = null;
@@ -49,6 +54,74 @@ export class PostDetailComponent implements OnInit {
   relateposts: Post[] = [];
   isLoggedIn = false;
   defaultImage = 'assets/images/default-avatar.jpg';
+  safeContent: SafeHtml | null = null;
+  currentUserId: number | null = null;
+  isPostOwner: boolean = false;
+  editingComment: Comment | null = null;
+  editingContent: string = '';
+
+  availableLanguages = [
+    { name: 'English', locale: 'en', flag: 'https://flagcdn.com/w20/us.png' },
+    { name: 'Tiếng Việt', locale: 'vi', flag: 'https://flagcdn.com/w20/vn.png' },
+    { name: 'Français', locale: 'fr', flag: 'https://flagcdn.com/w20/fr.png' },
+    { name: 'Deutsch', locale: 'de', flag: 'https://flagcdn.com/w20/de.png' },
+    { name: 'Español', locale: 'es', flag: 'https://flagcdn.com/w20/es.png' },
+    { name: '中文', locale: 'zh', flag: 'https://flagcdn.com/w20/cn.png' },
+    { name: '日本語', locale: 'ja', flag: 'https://flagcdn.com/w20/jp.png' },
+    { name: '한국어', locale: 'ko', flag: 'https://flagcdn.com/w20/kr.png' }
+  ].sort((a, b) => a.name.localeCompare(b.name));
+  selectedLanguage: string = 'vi';
+  showDropdown: boolean = false;
+
+  onLanguageChange(event: any) {
+    const lang = event?.target?.value || this.selectedLanguage;
+    this.selectedLanguage = lang;
+    if (!this.post?.id) return;
+    // Xác định locale gốc từ post.postTranslateLanguage (is_original=true)
+    let originalLocale = 'vi';
+    if (this.originalPost && Array.isArray(this.originalPost.postTranslateLanguage)) {
+      const originalTrans = this.originalPost.postTranslateLanguage.find(t => t.is_original && t.language && t.language.locale);
+      if (originalTrans) {
+        originalLocale = originalTrans.language.locale;
+      }
+    }
+    // Nếu là ngôn ngữ gốc thì revert về post gốc
+    if (lang === originalLocale) {
+      if (this.originalPost) {
+        this.post = { ...this.originalPost };
+        if (this.post.content) {
+          const clean = DOMPurify.sanitize(this.post.content, {ALLOWED_TAGS: ['b', 'i', 'em', 'strong', 'a', 'ul', 'ol', 'li', 'br', 'p', 'h1', 'h2', 'h3', 'h4', 'h5', 'h6', 'blockquote', 'pre', 'code', 'img'], ALLOWED_ATTR: ['href', 'src', 'alt', 'title', 'target', 'rel']});
+          this.safeContent = this.sanitizer.bypassSecurityTrustHtml(clean);
+        }
+      }
+      return;
+    }
+    // Gọi API lấy bản dịch
+    this.languagesService.getPostTranslation(this.post.id, lang).subscribe({
+      next: (res: any) => {
+        // Nếu có bản dịch thì cập nhật lại post
+        if (res && res.title) {
+          this.post = {
+            ...this.post!,
+            title: res.title,
+            description: res.description,
+            content: res.content
+          };
+          this.safeContent = this.sanitizer.bypassSecurityTrustHtml(DOMPurify.sanitize(res.content, {ALLOWED_TAGS: ['b', 'i', 'em', 'strong', 'a', 'ul', 'ol', 'li', 'br', 'p', 'h1', 'h2', 'h3', 'h4', 'h5', 'h6', 'blockquote', 'pre', 'code', 'img'], ALLOWED_ATTR: ['href', 'src', 'alt', 'title', 'target', 'rel']}));
+        }
+      },
+      error: () => {
+        // Nếu lỗi thì giữ nguyên post hiện tại
+      }
+    });
+  }
+
+  get selectedLanguageObj() {
+    return this.availableLanguages.find(l => l.locale === this.selectedLanguage);
+  }
+  get selectedLanguageFlag() {
+    return this.selectedLanguageObj?.flag;
+  }
 
   constructor(
     private route: ActivatedRoute,
@@ -56,7 +129,10 @@ export class PostDetailComponent implements OnInit {
     private commentsService: CommentsService,
     private authService: AuthService,
     private toastr: ToastrService,
-    private cloudinaryService: CloudinaryService
+    private cloudinaryService: CloudinaryService,
+    private sanitizer: DomSanitizer,
+    private router: Router,
+    private languagesService: LanguagesService
   ) {}
 
   private isNumeric(value: string): boolean {
@@ -105,12 +181,23 @@ export class PostDetailComponent implements OnInit {
       this.authService.isLoggedIn$.subscribe(loggedIn => {
         this.isLoggedIn = loggedIn;
       });
+
+      this.authService.currentUser$.subscribe(user => {
+        this.currentUserId = user?.id || null;
+        this.isPostOwner = this.currentUserId === this.post?.author?.id;
+      });
     });
   }
 
   private _processPostData(data: any): void {
     console.log('_processPostData called with:', data);
     this.post = data.post;
+    this.originalPost = data.post ? { ...data.post } : null; // Lưu lại post gốc
+    if (this.post && this.post.content) {
+      // Làm sạch nội dung bằng DOMPurify
+      const clean = DOMPurify.sanitize(this.post.content, {ALLOWED_TAGS: ['b', 'i', 'em', 'strong', 'a', 'ul', 'ol', 'li', 'br', 'p', 'h1', 'h2', 'h3', 'h4', 'h5', 'h6', 'blockquote', 'pre', 'code', 'img'], ALLOWED_ATTR: ['href', 'src', 'alt', 'title', 'target', 'rel']});
+      this.safeContent = this.sanitizer.bypassSecurityTrustHtml(clean);
+    }
     this.loading = false;
     if (this.post?.category_id) {
       this.loadRelatedPosts(this.post.category_id);
@@ -119,7 +206,9 @@ export class PostDetailComponent implements OnInit {
       this.loadComments(this.post.id);
     }
   }
-
+  getSanitizedContent(content: string): SafeHtml {
+    return this.sanitizer.bypassSecurityTrustHtml(content);
+  }
   private _handlePostError(error: Error): void {
     console.log('_handlePostError called with:', error);
     this.error = 'Failed to load post. Please try again later.';
@@ -140,12 +229,13 @@ export class PostDetailComponent implements OnInit {
 
   loadComments(postId: number): void {
     this.commentsService.getCommentsByPost(postId).subscribe({
-      next: (response: { comments: any[], pagination: any }) => {
-        this.comments = (response.comments || []).map(item => {
+      next: (response: any) => {
+        this.comments = (response.data?.comments || []).map((item: any) => {
           const comment = item.comment || item;
           if (item.user && comment) {
             comment.user = item.user;
           }
+          (comment as any).showMenu = false;
           return comment;
         });
       },
@@ -167,8 +257,20 @@ export class PostDetailComponent implements OnInit {
       return;
     }
 
-    if (!this.comments) {
-      this.comments = [];
+    if (this.editingComment) {
+      // Đang sửa comment
+      this.commentsService.updateComment(this.editingComment.id, { content: this.editingContent }).subscribe({
+        next: () => {
+          this.toastr.success('Cập nhật bình luận thành công!');
+          this.editingComment = null;
+          this.newComment = '';
+          this.loadComments(this.post!.id);
+        },
+        error: () => {
+          this.toastr.error('Cập nhật bình luận thất bại!');
+        }
+      });
+      return;
     }
 
     this.authService.currentUser$.pipe(
@@ -176,7 +278,6 @@ export class PostDetailComponent implements OnInit {
       switchMap(currentUser => {
         if (!currentUser || !currentUser.id) {
           this.toastr.error('Please log in to post a comment.');
-          console.error('Attempted to post comment without a logged-in user or user ID.');
           return throwError(() => new Error('User not logged in'));
         }
 
@@ -186,19 +287,7 @@ export class PostDetailComponent implements OnInit {
           user_id: currentUser.id,
         };
 
-        return this.commentsService.addComment(this.post!.id, commentDto).pipe(
-          map((response: { message: string, comment: Comment, user: User }) => {
-            const newCommentWithUser: Comment = {
-                ...response.comment,
-                user: response.user
-            };
-            return newCommentWithUser;
-          }),
-          catchError((error: any) => {
-            console.error('Error from addComment service:', error);
-            return throwError(() => error);
-          })
-        );
+        return this.commentsService.addComment(this.post!.id, commentDto);
       }),
       catchError((error: any) => {
         if (error.message !== 'User not logged in') {
@@ -207,21 +296,21 @@ export class PostDetailComponent implements OnInit {
         return of(null);
       })
     ).subscribe({
-      next: (comment: Comment | null) => {
-        if (comment) {
-            this.comments.unshift(comment);
-            this.newComment = '';
-            if (this.post) {
-              this.post.comments_count = (this.post.comments_count || 0) + 1;
-            }
-            this.toastr.success('Comment posted successfully!');
+      next: (response: any) => {
+        if (response) {
+          this.newComment = '';
+          this.toastr.success('Comment posted successfully!');
+          // Gọi lại loadComments để lấy danh sách comment mới nhất từ DB
+          if (this.post) {
+            this.loadComments(this.post.id);
+          }
         }
       }
     });
   }
 
   getImageUrl(url: string | undefined): string {
-    if (!url) return 'assets/images/default-avatar.jpg';
+    if (!url) return 'https://res.cloudinary.com/dejapatma/image/upload/v1751512040/uploads/g8gxuakcviyjxrwqsz52.jpg';
     
     // Handle Gravatar URLs for emails
     if (url.includes('@') && url.includes('.')) { // Simple check for email format
@@ -251,7 +340,7 @@ export class PostDetailComponent implements OnInit {
     }
     
     // Default fallback
-    return 'assets/images/default-avatar.jpg';
+    return 'https://res.cloudinary.com/dejapatma/image/upload/v1751512040/uploads/g8gxuakcviyjxrwqsz52.jpg';
   }
 
   getThumbnailUrl(post: Post): string {
@@ -280,11 +369,91 @@ export class PostDetailComponent implements OnInit {
       console.error('Error getting thumbnail URL:', error);
     }
     
-    return 'assets/images/default-avatar.jpg';
+    return 'https://res.cloudinary.com/dejapatma/image/upload/v1751512040/uploads/g8gxuakcviyjxrwqsz52.jpg';
   }
 
   handleImageError(event: Event): void {
     const imgElement = event.target as HTMLImageElement;
     imgElement.src = this.defaultImage;
+  }
+
+  goToProfile(authorId: number | string | undefined) {
+    console.log('authorId:', authorId, 'currentUserId:', localStorage.getItem('userId'));
+    if (!authorId) return;
+    const currentUserId = localStorage.getItem('userId');
+    if (String(authorId) === String(currentUserId)) {
+      this.router.navigate(['/profile/profile-user']);
+    } else {
+      this.router.navigate(['/profile/profile-settings', String(authorId)]);
+    }
+  }
+
+  isCommentOwner(comment: Comment): boolean {
+    return this.currentUserId === comment.user_id;
+  }
+
+  deleteComment(comment: Comment) {
+    // Kiểm tra id trước khi gọi API
+    this.commentsService.deleteComment(comment.id).subscribe({
+      next: () => {
+        this.toastr.success('Đã xóa bình luận!');
+        this.loadComments(this.post!.id);
+      },
+      error: () => {
+        this.toastr.error('Xóa bình luận thất bại!');
+      }
+    });
+  }
+
+  editComment(comment: Comment) {
+    this.editingComment = { ...comment };
+    this.editingContent = comment.content;
+  }
+
+  saveEditComment() {
+    if (!this.editingComment) return;
+    this.commentsService.updateComment(this.editingComment.id, { content: this.editingContent }).subscribe({
+      next: () => {
+        this.toastr.success('Cập nhật bình luận thành công!');
+        this.editingComment = null;
+        this.editingContent = '';
+        this.loadComments(this.post!.id);
+      },
+      error: () => {
+        this.toastr.error('Cập nhật bình luận thất bại!');
+      }
+    });
+  }
+
+  cancelEdit() {
+    this.editingComment = null;
+    this.editingContent = '';
+  }
+
+  toggleMenu(comment: any, forceClose: boolean = false) {
+    // Đóng tất cả menu khác
+    this.comments.forEach(c => { if (c !== comment) c.showMenu = false; });
+    if (forceClose) {
+      comment.showMenu = false;
+    } else {
+      comment.showMenu = !comment.showMenu;
+    }
+  }
+
+  async confirmDeleteComment(comment: Comment) {
+    const result = await Swal.fire({
+      title: 'Bạn có chắc muốn xóa bình luận này?',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Xóa',
+      cancelButtonText: 'Hủy',
+      confirmButtonColor: '#d33',
+      cancelButtonColor: '#3085d6',
+    });
+    if (result.isConfirmed) {
+      this.deleteComment(comment);
+    }
+    // Đóng menu
+    this.toggleMenu(comment, true);
   }
 }
