@@ -4,6 +4,6 @@ import { ProfileSettingsComponent } from './profile-settings/profile-settings.co
 
 export const routes: Routes = [
   { path: 'profile-user', component: ProfileUserComponent },
-  { path: 'profile', component: ProfileSettingsComponent },
+  { path: 'profile-settings/:id', component: ProfileSettingsComponent },
   { path: '', redirectTo: 'profile-user', pathMatch: 'full' }
 ];
